requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
django-filter==23.3
django-storages==1.14.2
boto3==1.28.62 